implementation/d05/requirements.txt
Django==2.2.7
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
